Fix type conversion errors

+++ call.go
@@ -16,7 +16,7 @@ func (calls *TGCalls) Mute(chatId int64) (int, error) {
 	if err != nil {
 		return Err, err
 	}
-	return result.(int), nil
+	return int(result.(int64)), nil
 }
 
 func (calls *TGCalls) Unmute(chatId int64) (int, error) {
@@ -27,7 +27,7 @@ func (calls *TGCalls) Unmute(chatId int64) (int, error) {
 	if err != nil {
 		return Err, err
 	}
-	return result.(int), nil
+	return int(result.(int64)), nil
 }
 
 func (calls *TGCalls) Pause(chatId int64) (int, error) {
@@ -38,7 +38,7 @@ func (calls *TGCalls) Pause(chatId int64) (int, error) {
 	if err != nil {
 		return Err, err
 	}
-	return result.(int), nil
+	return int(result.(int64)), nil
 }
 
 func (calls *TGCalls) Resume(chatId int64) (int, error) {
@@ -49,7 +49,7 @@ func (calls *TGCalls) Resume(chatId int64) (int, error) {
 	if err != nil {
 		return Err, err
 	}
-	return result.(int), nil
+	return int(result.(int64)), nil
 }
 
 func (calls *TGCalls) Stop(chatId int64) (int, error) {
@@ -60,5 +60,5 @@ func (calls *TGCalls) Stop(chatId int64) (int, error) {
 	if err != nil {
 		return Err, err
 	}
-	return result.(int), nil
+	return int(result.(int64)), nil
 }
